chore: refactor constant objects
