+++ pom.xml
@@ -23,7 +23,6 @@
                 <configuration>
                     <excludes>
                         <exclude>**/constant/**</exclude>
-                        <exclude>**/utils/**</exclude>
 
                         <exclude>**/client/openapi/OpenApiStub*</exclude>
                         <exclude>**/client/Client*</exclude>
@@ -150,13 +149,13 @@
 
         <!-- Logging -->
         <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.apache.logging.log4j</groupId>
             <artifactId>log4j-api</artifactId>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
         </dependency>
 
         <!-- Kotlin Atomic -->
