+++ src/main/kotlin/com/expediagroup/sdk/core/config/Config.kt
@@ -17,7 +17,7 @@ package com.expediagroup.sdk.core.config
 
 import com.expediagroup.sdk.core.config.provider.ConfigurationData
 import com.expediagroup.sdk.core.model.exception.ConfigurationException
-import com.expediagroup.sdk.core.utils.ExceptionMessageProvider.configurationUnknown
+import com.expediagroup.sdk.core.constant.ExceptionMessageProvider.configurationUnknown
 
 /**
  * @property configurationData deserialized configuration data which needs to be parsed
